refactor(python-sdk): move v3 handlers
Split `main.py` into individual handler modules under `handlers`.
Extract shared API client initialization.

Related-Task: INTER-1880
